Preparatory refactor: make /account's cards work for a named person
Behaviour identical, screenshots identical, no feature. This is the "make
the change easy" step for #104: `/account` was 1,254 lines in which every
card read the session directly, so the per-child page in #106 would have
had to be a copy-paste of a thousand of them.

Two things change shape.

`src/lib/household.ts` is new, and holds the rule about who may act for
whom in ONE place:

  * `assertActingFor(admin, callerId, targetId)` — the single gate. Passes
    for the caller themself or one of their own dependants. Refuses
    somebody else's dependant, an unknown target, and a caller who is
    themselves a dependant acting for anybody at all. That last case is
    the one-level rule, which the migration deliberately left to the
    application because a depth check in SQL needs a trigger.
  * `listHousehold(admin, userId)` — the account holder plus dependants.
  * `contactUserIdFor(profile)` — pure: a dependant's guardian, else the
    person themself.

Every refusal uses the same sentence whatever the cause, so the gate
cannot be used to probe whether a uuid is a real person at the club.

The six cards move to `src/components/site/account/` and take a `userId`
prop, and the four server functions they read and write through
(`getMyProfile`, `updateMyProfile`, `listMyWaivers`,
`getCodeOfConductSigner`) each gain an optional target that goes through
the gate. Absent still means the caller, so `/waiver` and
`/update-password` are untouched; `/account` passes the signed-in
person's own id. What the cards preserve is deliberate: the page still
fetches the profile once and shares it, each card's `stored` memo is
still keyed on its own fields rather than the profile object's identity,
and `loadFailed` is still tracked apart from `profile === null`.

`getCodeOfConductSigner` accepts a target only from a live session, never
from an emailed link: signing a waiver is public and hands back a
code-of-conduct token, so a token proves an address and must never prove
the right to read a household.

Two things this does NOT do, both noted in code:

  * `getWaiverPdfUrl` is unchanged and still scoped by RLS to the
    caller's own waivers, so a guardian will be able to list a
    dependant's waivers and not open one. Widening it is a policy change,
    which is a migration, and #104 carries none.
  * `media_consent_updated_by` stamps whoever clicked, so a guardian's
    answer for a dependant records the guardian. That is the honest
    record, and it means the person page reads it as somebody else's
    change rather than the dependant's own.

`docs/database.md` picks up `profiles.guardian_user_id` here, per
docs/database-changes.md: the migration is already live (#103/#108), and
this is the code change that first reads the column.

Tests: new `src/lib/household.test.ts` (17) covers the gate for self, own
dependant, someone else's dependant, a dependant acting as anyone, the
one-level rule against a deliberately bad row, an unknown target getting
the same words, and a failed read surfacing rather than being read as
allowed or denied; plus `contactUserIdFor` and `listHousehold`.
`account.test.tsx` follows the extraction and pins that every read names
the person by id. `validation.test.ts` covers the new `userId` key,
including that a patch which only names a person is still nothing to
update.

Verified: lint, typecheck, 2603 tests and build all green.
