+++ src/lib/profile.functions.ts
@@ -1,4 +1,7 @@
-// The details a signed-in person maintains about themselves, from `/account`.
+// The details a signed-in person maintains about the people on their account,
+// from `/account`. With no `userId` that is themselves, which is every caller
+// today; with one it is checked by `assertActingFor` (`src/lib/household.ts`),
+// so a guardian can reach a dependant and nobody else can reach anyone.
 //
 // This is the self-serve write path onto `profiles`. It covers what somebody
 // goes by, their kit sizes, and how the club reaches them. It deliberately
@@ -16,12 +19,25 @@
 import { createServerFn } from "@tanstack/react-start";
 import { requireSupabaseAuth } from "@/integrations/supabase/auth-middleware";
 import { updateMyProfileSchema } from "@/lib/validation";
-import type { UpdateMyProfileInput } from "@/lib/validation";
+import type { UpdateMyProfileFields } from "@/lib/validation";
+import { assertActingFor } from "@/lib/household";
 
 export const updateMyProfile = createServerFn({ method: "POST" })
   .middleware([requireSupabaseAuth])
   .inputValidator((d: unknown) => updateMyProfileSchema.parse(d))
   .handler(async ({ data, context }) => {
+    // `userId` names WHO this is about and is not a column, so it comes off
+    // before anything else touches the patch. Absent means the caller.
+    const { userId: target, ...fields } = data;
+
+    const { supabaseAdmin } = await import("@/integrations/supabase/client.server");
+    // The gate runs first, and only when a target was named: with none, this is
+    // the self-serve path it has always been. Before anything else, so that a
+    // patch which turns out to be empty is still refused rather than answered
+    // with a cheerful no-op that tells the caller their target was acceptable.
+    if (target) await assertActingFor(supabaseAdmin, context.userId, target);
+    const subjectId = target ?? context.userId;
+
     // Each card on /account sends only its own keys, so drop the absent ones:
     // `undefined` means "leave it alone", while an explicit `null` on a
     // nullable field means "clear it" and must survive to the UPDATE.
@@ -29,19 +45,23 @@ export const updateMyProfile = createServerFn({ method: "POST" })
     // still checked against the generated column list, so a key added to the
     // schema that is not a profiles column fails the typecheck here.
     const patch = Object.fromEntries(
-      Object.entries(data).filter(([, value]) => value !== undefined),
-    ) as UpdateMyProfileInput;
+      Object.entries(fields).filter(([, value]) => value !== undefined),
+    ) as UpdateMyProfileFields;
     // The schema's refine already rejects an empty patch; this keeps a future
     // caller from turning that into a pointless round trip.
     if (Object.keys(patch).length === 0) return { ok: true as const, fields: [] as string[] };
 
-    const { supabaseAdmin } = await import("@/integrations/supabase/client.server");
     const now = new Date().toISOString();
     // Media consent carries its own provenance, because the person page has to
     // tell a withdrawal the member made apart from one a manager recorded on
     // their behalf. Stamping the member's own id here is what makes the first
     // case distinguishable at all: `media_consent_updated_by === user_id` means
     // they set it themselves.
+    //
+    // It stamps whoever actually clicked, so a guardian answering for a
+    // dependant records the GUARDIAN's id, not the dependant's. That is the
+    // honest record, and it means the "they set it themselves" test above reads
+    // a guardian's answer as somebody else's, which is exactly what it is.
     const provenance =
       patch.media_consent === undefined
         ? {}
@@ -49,7 +69,7 @@ export const updateMyProfile = createServerFn({ method: "POST" })
     const { data: updated, error } = await supabaseAdmin
       .from("profiles")
       .update({ ...patch, ...provenance, updated_at: now })
-      .eq("user_id", context.userId)
+      .eq("user_id", subjectId)
       .select("user_id");
     if (error) throw new Error(error.message);
     // PostgREST reports no error when the filter matched nothing, so without
